removed virtwrap/agent dependency in launcher-common
