+++ cmd/fake-cmd-server/BUILD.bazel
@@ -8,8 +8,8 @@ go_library(
     deps = [
         "//pkg/virt-handler/cmd-client:go_default_library",
         "//pkg/virt-launcher-common:go_default_library",
+        "//pkg/virt-launcher-common/agent:go_default_library",
         "//pkg/virt-launcher-common/cmd-server:go_default_library",
-        "//pkg/virt-launcher/virtwrap/agent:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/github.com/spf13/pflag:go_default_library",
         "//vendor/go.uber.org/mock/gomock:go_default_library",
