+++ pkg/virt-launcher-common/agent/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["types.go"],
+    importpath = "kubevirt.io/kubevirt/pkg/virt-launcher-common/agent",
+    visibility = ["//visibility:public"],
+)
