+++ pkg/virt-launcher-common/cmd-server/BUILD.bazel
@@ -14,7 +14,7 @@ go_library(
         "//pkg/util/net/grpc:go_default_library",
         "//pkg/virt-handler/cmd-client:go_default_library",
         "//pkg/virt-launcher-common:go_default_library",
-        "//pkg/virt-launcher/virtwrap/agent:go_default_library",
+        "//pkg/virt-launcher-common/agent:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/json:go_default_library",
@@ -34,9 +34,9 @@ go_test(
         "//pkg/handler-launcher-com/cmd/v1:go_default_library",
         "//pkg/virt-handler/cmd-client:go_default_library",
         "//pkg/virt-launcher-common:go_default_library",
+        "//pkg/virt-launcher-common/agent:go_default_library",
         "//pkg/virt-launcher-common/api:go_default_library",
         "//pkg/virt-launcher-common/stats:go_default_library",
-        "//pkg/virt-launcher/virtwrap/agent:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
