+++ pkg/virt-launcher/virtwrap/agent/BUILD.bazel
@@ -5,5 +5,8 @@ go_library(
     srcs = ["exec.go"],
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/agent",
     visibility = ["//visibility:public"],
-    deps = ["//pkg/virt-launcher/virtwrap/cli:go_default_library"],
+    deps = [
+        "//pkg/virt-launcher-common/agent:go_default_library",
+        "//pkg/virt-launcher/virtwrap/cli:go_default_library",
+    ],
 )
